Don't rebuild the boot image every time
Fugu uses this for 2ndbootloader.bin, which is included in the boot and
recovery images. Since the size check destination never existed, it was
marked dirty and was re-run on every build. Change the size check to use
an empty file so that it's only run when the built image updates.

Change-Id: I0b1cef1ef0a825b1a1f89b0e90dda8706fbf7bf6